+++ backend/src/main/java/com/techeer/carpool/domain/post/controller/PostController.java
@@ -3,7 +3,10 @@ package com.techeer.carpool.domain.post.controller;
 import com.techeer.carpool.domain.post.dto.PostCreateRequest;
 import com.techeer.carpool.domain.post.dto.PostResponse;
 import com.techeer.carpool.domain.post.dto.PostUpdateRequest;
-import com.techeer.carpool.domain.post.service.PostService;
+import com.techeer.carpool.domain.post.service.PostCreateService;
+import com.techeer.carpool.domain.post.service.PostDeleteService;
+import com.techeer.carpool.domain.post.service.PostReadService;
+import com.techeer.carpool.domain.post.service.PostUpdateService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,33 +19,36 @@ import java.util.List;
 @RequiredArgsConstructor
 public class PostController {
 
-    private final PostService postService;
+    private final PostCreateService postCreateService;
+    private final PostReadService postReadService;
+    private final PostUpdateService postUpdateService;
+    private final PostDeleteService postDeleteService;
 
     @PostMapping
     public ResponseEntity<PostResponse> createPost(@RequestBody PostCreateRequest request) {
-        return ResponseEntity.status(HttpStatus.CREATED).body(postService.createPost(request));
+        return ResponseEntity.status(HttpStatus.CREATED).body(postCreateService.createPost(request));
     }
 
     @GetMapping
     public ResponseEntity<List<PostResponse>> getAllPosts() {
-        return ResponseEntity.ok(postService.getAllPosts());
+        return ResponseEntity.ok(postReadService.getAllPosts());
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<PostResponse> getPostById(@PathVariable Long id) {
-        return ResponseEntity.ok(postService.getPostById(id));
+        return ResponseEntity.ok(postReadService.getPostById(id));
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<PostResponse> updatePost(
             @PathVariable Long id,
             @RequestBody PostUpdateRequest request) {
-        return ResponseEntity.ok(postService.updatePost(id, request));
+        return ResponseEntity.ok(postUpdateService.updatePost(id, request));
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deletePost(@PathVariable Long id) {
-        postService.deletePost(id);
+        postDeleteService.deletePost(id);
         return ResponseEntity.noContent().build();
     }
 }
