+++ backend/src/main/java/com/techeer/carpool/domain/post/service/PostCreateService.java
@@ -2,22 +2,15 @@ package com.techeer.carpool.domain.post.service;
 
 import com.techeer.carpool.domain.post.dto.PostCreateRequest;
 import com.techeer.carpool.domain.post.dto.PostResponse;
-import com.techeer.carpool.domain.post.dto.PostUpdateRequest;
 import com.techeer.carpool.domain.post.entity.Post;
 import com.techeer.carpool.domain.post.repository.PostRepository;
-import com.techeer.carpool.global.exception.CarpoolException;
-import com.techeer.carpool.global.exception.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class PostService {
+public class PostCreateService {
 
     private final PostRepository postRepository;
 
@@ -39,33 +32,5 @@ public class PostService {
                 .build();
 
         return PostResponse.from(postRepository.save(post));
-    }
-
-    public List<PostResponse> getAllPosts() {
-        return postRepository.findByDeletedFalseOrderByCreatedAtDesc()
-                .stream()
-                .map(PostResponse::from)
-                .collect(Collectors.toList());
-    }
-
-    public PostResponse getPostById(Long id) {
-        Post post = postRepository.findByIdAndDeletedFalse(id)
-                .orElseThrow(() -> new CarpoolException(ErrorCode.POST_NOT_FOUND));
-        return PostResponse.from(post);
-    }
-
-    @Transactional
-    public PostResponse updatePost(Long id, PostUpdateRequest request) {
-        Post post = postRepository.findByIdAndDeletedFalse(id)
-                .orElseThrow(() -> new CarpoolException(ErrorCode.POST_NOT_FOUND));
-        post.updateFrom(request);
-        return PostResponse.from(post);
-    }
-
-    @Transactional
-    public void deletePost(Long id) {
-        Post post = postRepository.findByIdAndDeletedFalse(id)
-                .orElseThrow(() -> new CarpoolException(ErrorCode.POST_NOT_FOUND));
-        post.delete();
     }
 }
