+++ backend/src/main/java/com/techeer/carpool/domain/post/service/PostReadService.java
@@ -1,0 +1,34 @@
+package com.techeer.carpool.domain.post.service;
+
+import com.techeer.carpool.domain.post.dto.PostResponse;
+import com.techeer.carpool.domain.post.entity.Post;
+import com.techeer.carpool.domain.post.repository.PostRepository;
+import com.techeer.carpool.global.exception.CarpoolException;
+import com.techeer.carpool.global.exception.ErrorCode;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class PostReadService {
+
+    private final PostRepository postRepository;
+
+    public List<PostResponse> getAllPosts() {
+        return postRepository.findByDeletedFalseOrderByCreatedAtDesc()
+                .stream()
+                .map(PostResponse::from)
+                .collect(Collectors.toList());
+    }
+
+    public PostResponse getPostById(Long id) {
+        Post post = postRepository.findByIdAndDeletedFalse(id)
+                .orElseThrow(() -> new CarpoolException(ErrorCode.POST_NOT_FOUND));
+        return PostResponse.from(post);
+    }
+}
