+++ backend/src/main/java/com/techeer/carpool/domain/post/service/PostUpdateService.java
@@ -1,0 +1,26 @@
+package com.techeer.carpool.domain.post.service;
+
+import com.techeer.carpool.domain.post.dto.PostResponse;
+import com.techeer.carpool.domain.post.dto.PostUpdateRequest;
+import com.techeer.carpool.domain.post.entity.Post;
+import com.techeer.carpool.domain.post.repository.PostRepository;
+import com.techeer.carpool.global.exception.CarpoolException;
+import com.techeer.carpool.global.exception.ErrorCode;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class PostUpdateService {
+
+    private final PostRepository postRepository;
+
+    @Transactional
+    public PostResponse updatePost(Long id, PostUpdateRequest request) {
+        Post post = postRepository.findByIdAndDeletedFalse(id)
+                .orElseThrow(() -> new CarpoolException(ErrorCode.POST_NOT_FOUND));
+        post.updateFrom(request);
+        return PostResponse.from(post);
+    }
+}
